requirements.txt
streamlit
tensorflow
opencv-python-headless
matplotlib
scikit-learn
Pillow